Split tests into seperate file
